+++ griptape/artifacts/csv_row_artifact.py
@@ -5,16 +5,19 @@ import io
 
 from attrs import define, field
 
-from griptape.artifacts import BaseArtifact, TextArtifact
+from griptape.artifacts import BaseArtifact, BaseTextArtifact
 
 
 @define
-class CsvRowArtifact(TextArtifact):
+class CsvRowArtifact(BaseTextArtifact):
     value: dict[str, str] = field(converter=BaseArtifact.value_to_dict, metadata={"serializable": True})
     delimiter: str = field(default=",", kw_only=True, metadata={"serializable": True})
 
     def __add__(self, other: BaseArtifact) -> CsvRowArtifact:
         return CsvRowArtifact(self.value | other.value)
+
+    def __bool__(self) -> bool:
+        return len(self) > 0
 
     def to_text(self) -> str:
         with io.StringIO() as csvfile:
@@ -28,6 +31,3 @@ class CsvRowArtifact(TextArtifact):
             writer.writerow(self.value)
 
             return csvfile.getvalue().strip()
-
-    def __bool__(self) -> bool:
-        return len(self) > 0